+++ kurtosis-module/impl/participant_network/participant_network.go
@@ -32,10 +32,14 @@ const (
 	elClientMineWaiterMaxNumRetriesPerNode = uint32(120)
 	elClientMineWaiterTimeBetweenRetries = 5 * time.Second
 
+	// The time that the CL genesis generation step takes to complete, based off what we've seen
+	clGenesisDataGenerationTime = 2 * time.Minute
+
 	// Each CL node takes about this time to start up and start processing blocks, so when we create the CL
 	//  genesis data we need to set the genesis timestamp in the future so that nodes don't miss important slots
 	// (e.g. Altair fork)
-	clNodeStartupTime = 70 * time.Second
+	// TODO Make this client-specific (currently this is Nimbus)
+	clNodeStartupTime = 45 * time.Second
 )
 
 // To get clients to start as bootnodes, we pass in these values when starting them
@@ -172,8 +176,10 @@ func LaunchParticipantNetwork(
 	// We create the CL genesis data after the EL network is ready so that the CL genesis timestamp will be close
 	//  to the time the CL nodes are started
 	logrus.Info("Generating CL client genesis data...")
-	// Set the genesis timestamp in the future so we don't start running slots until all the nodes are up
-	clGenesisTimestamp := uint64(time.Now().Unix()) + uint64(numParticipants) * uint64(clNodeStartupTime.Seconds())
+	// Set the genesis timestamp in the future so we don't start running slots until all the CL nodes are up
+	clGenesisTimestamp := uint64(time.Now().Unix()) +
+		uint64(clGenesisDataGenerationTime.Seconds()) +
+		uint64(numParticipants) * uint64(clNodeStartupTime.Seconds())
 	clGenesisData, err := prelaunchDataGeneratorCtx.GenerateCLGenesisData(
 		clGenesisConfigTemplate,
 		clGenesisMnemonicsYmlTemplate,